test: split unit test for setup_name_resolver
