+++ test/unit/units/setup_host_network_test.py
@@ -1,30 +1,17 @@
-import io
 from unittest.mock import (
     patch, call, Mock
 )
 from pytest import raises
 
 from suse_migration_services.units.setup_host_network import (
-    main, has_host_resolv_setup
+    main
 )
 from suse_migration_services.exceptions import (
-    DistMigrationNameResolverException,
     DistMigrationHostNetworkException
 )
 
 
 class TestSetupHostNetwork(object):
-    @patch('suse_migration_services.logger.Logger.setup')
-    @patch('os.path.exists')
-    @patch('suse_migration_services.units.setup_host_network.Fstab')
-    def test_main_resolv_conf_not_present(
-        self, mock_Fstab, mock_os_path_exists, mock_logger_setup
-    ):
-        mock_os_path_exists.return_value = False
-        with raises(DistMigrationNameResolverException):
-            main()
-
-    @patch('suse_migration_services.units.setup_host_network.has_host_resolv_setup')
     @patch('suse_migration_services.logger.Logger.setup')
     @patch('suse_migration_services.command.Command.run')
     @patch('suse_migration_services.units.setup_host_network.Fstab')
@@ -32,17 +19,14 @@ class TestSetupHostNetwork(object):
     @patch('shutil.copy')
     def test_main_raises_on_host_network_activation(
         self, mock_shutil_copy, mock_os_path_exists,
-        mock_Fstab, mock_Command_run, mock_logger_setup, mock_resolv_empty
+        mock_Fstab, mock_Command_run, mock_logger_setup
     ):
         mock_os_path_exists.return_value = True
         mock_Command_run.side_effect = Exception
-        mock_resolv_empty.return_value = False
         with raises(DistMigrationHostNetworkException):
             main()
-        assert not mock_resolv_empty.called
         assert not mock_shutil_copy.called
 
-    @patch('suse_migration_services.units.setup_host_network.has_host_resolv_setup')
     @patch('suse_migration_services.logger.Logger.setup')
     @patch('suse_migration_services.command.Command.run')
     @patch('suse_migration_services.units.setup_host_network.Fstab')
@@ -52,7 +36,7 @@ class TestSetupHostNetwork(object):
     @patch('os.path.isfile')
     def test_main(
         self, mock_isfile, mock_glob, mock_shutil_copy, mock_os_path_exists,
-        mock_Fstab, mock_Command_run, mock_logger_setup, mock_empty_resolv
+        mock_Fstab, mock_Command_run, mock_logger_setup
     ):
         fstab = Mock()
         mock_Fstab.return_value = fstab
@@ -62,7 +46,6 @@ class TestSetupHostNetwork(object):
         ]
         mock_isfile.return_value = True
         mock_os_path_exists.return_value = True
-        mock_empty_resolv.return_value = True
         main()
 
         mock_glob.assert_called_once_with(
@@ -76,8 +59,7 @@ class TestSetupHostNetwork(object):
             call(
                 '/system-root/etc/sysconfig/network/dhcp',
                 '/etc/sysconfig/network'
-            ),
-            call('/system-root/etc/resolv.conf', '/etc/resolv.conf')
+            )
         ]
         assert mock_Command_run.call_args_list == [
             call(
@@ -99,90 +81,5 @@ class TestSetupHostNetwork(object):
             '/etc/sysconfig/network/providers'
         )
         fstab.export.assert_called_once_with(
-            '/etc/system-root.fstab'
-        )
-
-    def test_resolv_empty_with_info(self):
-        with patch(
-            'builtins.open',
-            return_value=io.StringIO('# foo\n# bar\nsearch domain'),
-            create=True
-        ):
-            assert has_host_resolv_setup('foo') is True
-
-    def test_resolv_empty_without_info(self):
-        with patch(
-            'suse_migration_services.units.setup_host_network.open',
-            return_value=io.StringIO('#foo\n#bar\n'),
-            create=True
-        ):
-            assert has_host_resolv_setup('foo') is False
-
-    @patch('suse_migration_services.units.setup_host_network.has_host_resolv_setup')
-    @patch('suse_migration_services.logger.Logger.setup')
-    @patch('suse_migration_services.command.Command.run')
-    @patch('suse_migration_services.units.setup_host_network.Fstab')
-    @patch('os.path.exists')
-    @patch('shutil.copy')
-    @patch('glob.glob')
-    @patch('os.path.isfile')
-    def test_main_bind_mount_resolv(
-        self, mock_isfile, mock_glob, mock_shutil_copy, mock_os_path_exists,
-        mock_Fstab, mock_Command_run, mock_logger_setup, mock_empty_resolv
-    ):
-        fstab = Mock()
-        mock_Fstab.return_value = fstab
-        mock_glob.return_value = [
-            '/system-root/etc/sysconfig/network/ifcfg-eth0',
-            '/system-root/etc/sysconfig/network/dhcp'
-        ]
-        mock_isfile.return_value = True
-        mock_os_path_exists.return_value = True
-        mock_empty_resolv.return_value = False
-        main()
-
-        mock_glob.assert_called_once_with(
-            '/system-root/etc/sysconfig/network/*'
-        )
-        assert mock_shutil_copy.call_args_list == [
-            call(
-                '/system-root/etc/sysconfig/network/ifcfg-eth0',
-                '/etc/sysconfig/network'
-            ),
-            call(
-                '/system-root/etc/sysconfig/network/dhcp',
-                '/etc/sysconfig/network'
-            )
-        ]
-        assert mock_Command_run.call_args_list == [
-            call(
-                [
-                    'mount', '--bind',
-                    '/system-root/etc/sysconfig/network/providers',
-                    '/etc/sysconfig/network/providers'
-                ]
-            ),
-            call(
-                ['systemctl', 'reload', 'network']
-            ),
-            call(
-                [
-                    'mount', '--bind',
-                    '/etc/resolv.conf',
-                    '/system-root/etc/resolv.conf'
-                ]
-            ),
-        ]
-        fstab.add_entry.assert_has_calls(
-            [
-                call('/system-root/etc/sysconfig/network/providers',
-                     '/etc/sysconfig/network/providers'),
-                call('/etc/resolv.conf', '/system-root/etc/resolv.conf'),
-            ]
-        )
-        fstab.read.assert_any_call(
-            '/etc/system-root.fstab'
-        )
-        fstab.export.assert_called_with(
             '/etc/system-root.fstab'
         )
